Working on the familial matching part of the pipeline.

+++ src/setupivoryfammatch.sh
@@ -6,18 +6,20 @@
 
 for i in 0 1 2 3 4 5
 do
-  rm -rf sub$i
-  mkdir sub$i
-  cp calculate_LRs.R LR_functions.R sub$i
-  mv new${i}.txt sub$i
-  mv old${i}.txt sub$i
-  mv ref${i}_fammatch.csv sub$i
-  cd sub$i
-  species='forest'
-  if [ $i -gt 1 ]
-  then
-    species='savannah'
+  if test -f "new$i"; then
+    rm -rf sub$i
+    mkdir sub$i
+    cp ../calculate_LRs.R ../LR_functions.R sub$i
+    mv new${i}.txt sub$i
+    mv old${i}.txt sub$i
+    mv ref${i}_fammatch.csv sub$i
+    cd sub$i
+    species='forest'
+    if [ $i -gt 1 ]
+    then
+      species='savannah'
+    fi
+    echo "Rscript calculate_LRs.R ${species} ref${i}_fammatch.csv old${i}.txt new${i}.txt" > runrscript.sh
+    cd .. 
   fi
-  echo "Rscript calculate_LRs.R ${species} ref${i}_fammatch.csv old${i}.txt new${i}.txt" > runrscript.sh
-  cd .. 
 done
